+++ pkg/api/controllers/routes.go
@@ -9,7 +9,7 @@ import (
 func (s *Server) initRoutes() {
 	logger.Debug("initializing Insprd server routes")
 	h := handler.NewHandler(
-		s.MemoryManager, s.op, s.auth,
+		s.MemoryManager, s.op, s.auth, s.BrokerManager,
 	)
 
 	ahandler := h.NewAppHandler()
@@ -23,6 +23,9 @@ func (s *Server) initRoutes() {
 
 	aliasHandler := h.NewAliasHandler()
 	s.Mux.Handle("/alias", rest.HandleCRUD(aliasHandler))
+
+	brokersHandler := h.NewBrokerHandler()
+	s.Mux.Handle("/brokers", brokersHandler.HandleGet())
 
 	s.Mux.Handle("/auth", h.TokenHandler().Validate(s.auth))
 	s.Mux.Handle("/refreshController", h.ControllerRefreshHandler())