+++ CHANGELOG.md
@@ -9,6 +9,18 @@
       `sidecars` pkg.
     - created the utils file containing small functions and mocks of functions
       to be, these functions should be able to be used by other sidecars.
+### #61 Story CORE-428 | CLI get installed brokers command
+- features:
+    - added a Brokers method and Brokers interface to controllers interface
+    - changed controller client to comply with new interface
+    - implemented a broker data interface (DI)
+    - changed insprd's server struct, http client, handler structure, and their mocks to include a BrokerManager
+    - created a handler to get broker data from insprd
+    - implemented cli command for getting broker data
+-  tests:
+    - unit tested handler
+    - integration test succeeded on cluster
+---
 
 ### #58  Story CORE-418 | Create Sidecar Factory
 - features:
@@ -20,11 +32,14 @@
     - tested previously mentioned methods
 - fixes:
     - moved broker management interfaces and structures to a more appropriate directory
+---
+
 ### #57 Tech CORE-426 | Review Permissions
 - features:
     - created functions in the UIDP client that make sure when creating a token, it’s permissions should be the same as the creator’s permissions (or have less permissions).
 - fixes:
     - all functions of the get command. For example, if you're trying to get channels, before calling the getDapp directly it will call the getChannels to check if the user have the right permissions.
+---
 
 ### #59 - Fix: CORE-310 | Automatic Helm chart tags update
 - fixes: